Removed database_models.go & cleaned up last model changes
