+++ project_test.go
@@ -112,3 +112,30 @@ func TestGetParamsWithOptionalEnvironment(t *testing.T) {
 		})
 	}
 }
+
+func TestDBProjectToResponseBuildDefParsing(t *testing.T) {
+	var testCases = []struct {
+		name      string
+		dbProject database.Project
+		want      interface{}
+	}{
+		{
+			name:      "no build def",
+			dbProject: database.Project{},
+			want:      nil,
+		},
+		{
+			name:      "with build def",
+			dbProject: database.Project{BuildDefinition: "myStage: moo"},
+			want: map[string]interface{}{
+				"myStage": "moo",
+			},
+		},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			resProject := dbProjectToResponse(tc.dbProject)
+			assert.Equal(t, tc.want, resProject.ParsedBuildDefinition)
+		})
+	}
+}
